fix: swagger routes and remove post video image upload link because of duplication
