+++ cms/djangoapps/contentstore/rest_api/v1/views/assets.py
@@ -2,7 +2,7 @@
 Public rest API endpoints for the CMS API Assets.
 """
 import logging
-from rest_framework.generics import RetrieveUpdateDestroyAPIView, CreateAPIView
+from rest_framework.generics import CreateAPIView, RetrieveAPIView, UpdateAPIView, DestroyAPIView
 from django.views.decorators.csrf import csrf_exempt
 from django.http import Http404
 
@@ -24,7 +24,7 @@ toggles = contentstore_toggles
 
 
 @view_auth_classes()
-class AssetsView(DeveloperErrorViewMixin, RetrieveUpdateDestroyAPIView, CreateAPIView):
+class AssetsReadPostView(DeveloperErrorViewMixin, CreateAPIView, RetrieveAPIView):
     """
     public rest API endpoints for the CMS API Assets.
     course_key: required argument, needed to authorize course authors and identify the asset.
@@ -44,16 +44,38 @@ class AssetsView(DeveloperErrorViewMixin, RetrieveUpdateDestroyAPIView, CreateAP
             raise Http404
         return super().dispatch(request, *args, **kwargs)
 
-    @course_author_access_required
-    @expect_json_in_class_view
-    def retrieve(self, request, course_key):  # pylint: disable=arguments-differ
-        return handle_assets(request, course_key.html_id())
-
     @csrf_exempt
     @course_author_access_required
     @validate_request_with_serializer
     def create(self, request, course_key):  # pylint: disable=arguments-differ
         return handle_assets(request, course_key.html_id())
+
+    @course_author_access_required
+    @expect_json_in_class_view
+    def retrieve(self, request, course_key):  # pylint: disable=arguments-differ
+        return handle_assets(request, course_key.html_id())
+
+
+@view_auth_classes()
+class AssetsView(DeveloperErrorViewMixin, UpdateAPIView, DestroyAPIView):
+    """
+    public rest API endpoints for the CMS API Assets.
+    course_key: required argument, needed to authorize course authors and identify the asset.
+    asset_key_string: required argument, needed to identify the asset.
+    """
+    serializer_class = AssetSerializer
+    parser_classes = (JSONParser, MultiPartParser, FormParser, TypedFileUploadParser)
+
+    def dispatch(self, request, *args, **kwargs):
+        # TODO: probably want to refactor this to a decorator.
+        """
+        The dispatch method of a View class handles HTTP requests in general
+        and calls other methods to handle specific HTTP methods.
+        We use this to raise a 404 if the content api is disabled.
+        """
+        if not toggles.use_studio_content_api():
+            raise Http404
+        return super().dispatch(request, *args, **kwargs)
 
     @course_author_access_required
     @expect_json_in_class_view
