+++ cms/djangoapps/contentstore/rest_api/v1/views/videos.py
@@ -143,29 +143,3 @@ class VideoFeaturesView(DeveloperErrorViewMixin, RetrieveAPIView):
     @csrf_exempt
     def retrieve(self, request):  # pylint: disable=arguments-differ
         return enabled_video_features(request)
-
-
-@view_auth_classes()
-class UploadLinkView(DeveloperErrorViewMixin, CreateAPIView):
-    """
-    public rest API endpoint providing a list of enabled video features.
-    """
-    serializer_class = VideoUploadSerializer
-
-    def dispatch(self, request, *args, **kwargs):
-        # TODO: probably want to refactor this to a decorator.
-        """
-        The dispatch method of a View class handles HTTP requests in general
-        and calls other methods to handle specific HTTP methods.
-        We use this to raise a 404 if the content api is disabled.
-        """
-        if not toggles.use_studio_content_api():
-            raise Http404
-        return super().dispatch(request, *args, **kwargs)
-
-    @csrf_exempt
-    @course_author_access_required
-    @expect_json_in_class_view
-    @validate_request_with_serializer
-    def create(self, request, course_key):  # pylint: disable=arguments-differ
-        return handle_videos(request, course_key.html_id())
